feat: 유저 정보 api 연결
Related to: #51
