+++ app/mypage/page.tsx
@@ -2,13 +2,12 @@
 
 import { useEffect } from 'react';
 import { useRouter } from 'next/navigation';
-import { Button } from '@nextui-org/button';
-import { Avatar } from '@nextui-org/avatar';
 import AddButon from '@/components/common/Button/AddButton';
 import Slider from 'react-slick';
 import { getCookie } from 'cookies-next';
 import { ACCESS_TOKEN } from '@/constants';
-import { PartyItem, CongressmanList } from './components';
+import { dehydrate, HydrationBoundary, useQueryClient } from '@tanstack/react-query';
+import { PartyItem, CongressmanList, UserInfo } from './components';
 
 const partyList = [
   {
@@ -111,6 +110,7 @@ const congressmanList = [
 export default function MyPage() {
   const router = useRouter();
   const accessToken = getCookie(ACCESS_TOKEN);
+  const queryClient = useQueryClient();
   const sliderSettings = {
     arrows: false,
     dots: false,
@@ -129,55 +129,40 @@ export default function MyPage() {
   if (!accessToken) return <div className="flex items-center justify-center h-full">회원 정보가 없습니다.</div>;
 
   return (
-    <div className="flex flex-col h-full gap-8">
-      <section className="flex items-center px-[30px] justify-between h-[200px] shadow-md rounded-xl bg-white pb-7">
-        <div className="flex items-center">
-          <Avatar src="/images/basicAvatar.png" className="w-[100px] h-[100px] mr-4" />
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <div className="flex flex-col h-full gap-8">
+        <UserInfo queryClient={queryClient} />
 
-          <div className="flex flex-col gap-3">
-            <p className="text-3xl font-semibold">홍길동</p>
-            <p className="text-[#999999] text-xs">abcd123@gmail.com</p>
+        <section className="pl-[30px] flex flex-col gap-6">
+          <div className="flex items-center justify-between pr-[30px]">
+            <p className="text-xl font-semibold">
+              팔로우한 정당 &middot;<span className="text-[#555555]"> 3</span>
+            </p>
+
+            <AddButon />
           </div>
-        </div>
 
-        <Button
-          radius="full"
-          size="sm"
-          variant="bordered"
-          className="h-8 bg-transparent border-1 border-[#E0E0E0] text-[#999999] ">
-          로그아웃
-        </Button>
-      </section>
+          <Slider {...sliderSettings}>
+            {partyList.map((party) => (
+              <PartyItem key={party.label} {...party} />
+            ))}
+          </Slider>
+        </section>
 
-      <section className="pl-[30px] flex flex-col gap-6">
-        <div className="flex items-center justify-between pr-[30px]">
-          <p className="text-xl font-semibold">
-            팔로우한 정당 &middot;<span className="text-[#555555]"> 3</span>
-          </p>
+        <hr className="mx-[30px] bg-[#E0E0E0]" />
 
-          <AddButon />
-        </div>
+        <section className="px-[30px] flex flex-col gap-6 pb-10">
+          <div className="flex items-center justify-between">
+            <p className="text-xl font-semibold">
+              팔로우한 의원 &middot;<span className="text-[#555555]"> 16</span>
+            </p>
 
-        <Slider {...sliderSettings}>
-          {partyList.map((party) => (
-            <PartyItem key={party.label} {...party} />
-          ))}
-        </Slider>
-      </section>
+            <AddButon />
+          </div>
 
-      <hr className="mx-[30px] bg-[#E0E0E0]" />
-
-      <section className="px-[30px] flex flex-col gap-6 pb-10">
-        <div className="flex items-center justify-between">
-          <p className="text-xl font-semibold">
-            팔로우한 의원 &middot;<span className="text-[#555555]"> 16</span>
-          </p>
-
-          <AddButon />
-        </div>
-
-        <CongressmanList congressmanList={congressmanList} />
-      </section>
-    </div>
+          <CongressmanList congressmanList={congressmanList} />
+        </section>
+      </div>
+    </HydrationBoundary>
   );
 }
